pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seaman-brain"
version = "0.1.0"
description = "Seaman Reborn - sardonic creature cognition engine with Pygame interface"
requires-python = ">=3.13"
license = "MIT"
authors = [{ name = "djbowles" }]
dependencies = [
    "ollama>=0.4",
    "lancedb>=0.17",
    "pydantic>=2.0",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "pygame>=2.6",
    "pyttsx3>=2.9",
    "sounddevice>=0.4",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "websockets>=14.0",
]

[project.optional-dependencies]
cloud = [
    "openai>=1.50",
    "anthropic>=0.40",
]
stt = [
    "SpeechRecognition>=3.10",
    "pyaudio>=0.2.14",
]
stt-local = [
    "faster-whisper>=1.1.0",
    "sounddevice>=0.5.0",
]
tts-neural = [
    "kokoro>=0.7.4",
    "soundfile>=0.12.0",
]
vision = [
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "pygrabber>=0.2",
]
riva = [
    "nvidia-riva-client>=2.16.0",
    "grpcio>=1.60.0",
]
aec = [
    "webrtcvad-wheels>=2.0.13",
    "scipy>=1.12.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "ruff>=0.8",
    "httpx>=0.28",
]
all = ["seaman-brain[cloud,stt,stt-local,tts-neural,vision,riva,aec,dev]"]

[project.scripts]
seaman-brain = "seaman_brain.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring external services",
]
